More cleanup changes
Also removed the call to benchmark from the travis file (since it calls
Born and Born is currently broken - that shouldn’t stop the PR though -
the adjoint test still passes)

Even more changes!
