+++ .travis.yml
@@ -26,4 +26,3 @@ script:
   - export PYTHONPATH=$PYTHONPATH:$PWD
   - flake8 .
   - py.test -vs tests/
-  - python benchmark.py
